requirements.txt
pyTelegramBotAPI==4.12.0
pynput==1.7.6
Pillow
psutil==5.9.5
